Phase 6 COMPLETE: Polish and end-to-end validation (63 tests)
Code quality:
- Fix all 45 clippy warnings (doc comments, collapsible ifs, casts, etc.)
- Add --help/-h CLI flag with full usage documentation
- Remove dead TcpBridge stub, rename _tcp_port → tcp_port
- Create lib.rs to expose modules for integration testing

Bug fixes:
- Fix MOV.L ERs, ERd decoder regression (commit 3bf52c9 broke 0F 9x/Bx-Fx)
  Firmware was halting at 0x0203FE during RAM test (MOV.L ER6, ER0)
- Fix boot_to_main_loop milestone check (was checking wrong marker key)

End-to-end integration test (16 tests):
- Boot real LS-50 firmware, validate full SCSI init + scan sequence
- Verify INQUIRY returns "Nikon LS-50 ED 1.02", scan produces 270KB image
- Test all SCSI opcodes: TUR, INQUIRY, REQUEST SENSE, RESERVE/RELEASE,
  MODE SENSE/SELECT, SET/GET WINDOW, SEND DIAGNOSTIC, READ, WRITE, SCAN,
  VENDOR C0/C1/E0/E1, plus illegal opcode error handling

+++ emulator/h8300h-core/src/execute.rs
@@ -1,7 +1,7 @@
-/// H8/300H instruction executor.
-///
-/// Each instruction modifies CPU registers, memory, and/or CCR flags
-/// according to the Hitachi H8/300H Programming Manual.
+//! H8/300H instruction executor.
+//!
+//! Each instruction modifies CPU registers, memory, and/or CCR flags
+//! according to the Hitachi H8/300H Programming Manual.
 
 use crate::cpu::*;
 use crate::decode::*;
@@ -95,7 +95,7 @@ pub fn execute(cpu: &mut Cpu, bus: &mut MemoryBus, insn: &Instruction, insn_pc: 
         Instruction::MulxuB(src, dst) => {
             let s = read_operand_b(cpu, bus, src) as u16;
             let d = read_operand_w(cpu, bus, dst);
-            let dl = (d & 0xFF) as u16;
+            let dl = d & 0xFF;
             let result = dl * s;
             write_operand_w(cpu, bus, dst, result);
             // No flags affected
@@ -104,7 +104,7 @@ pub fn execute(cpu: &mut Cpu, bus: &mut MemoryBus, insn: &Instruction, insn_pc: 
         Instruction::MulxuW(src, dst) => {
             let s = read_operand_w(cpu, bus, src) as u32;
             let d = read_operand_l(cpu, bus, dst);
-            let dl = (d & 0xFFFF) as u32;
+            let dl = d & 0xFFFF;
             let result = dl * s;
             write_operand_l(cpu, bus, dst, result);
             next_pc
